backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
xgboost==2.0.0
shap==0.43.0
SQLAlchemy==2.0.21
bcrypt==4.0.1
PyJWT==2.8.0
Flask-Cors==4.0.0